Added Save/Load parameters option and a README file and an RScript in the zip re
Added Save/Load parameters option. Also added two files in the zip
results file: one RScript that the user can run to create the plots
himself and a RData file with the necessary data for the script. Also
exposed the p threshold and other parameters. In addition, the user's IP
is no longer displayed.